Initial commit
A substantial portion of the HTML and CSS were taken from part 2 of the
project where I created a demo of the site features. At first I
just continued adding commits to the previously made repo, but this was
not what the directions said to do. Anyways, that is why so much code
already existed at this point.